make field examples reusable
