+++ crypto/finite-field-examples.ts
@@ -1,0 +1,27 @@
+/**
+ * This file contains some examples of finite fields, to be used for tests
+ */
+import { Fp, Fq, createField } from './finite_field.js';
+
+export { exampleFields };
+
+// some primes
+let pSmall = 101n;
+let pBabybear = (1n << 31n) - 1n;
+let pGoldilocks = (1n << 64n) - (1n << 32n) + 1n;
+let p25519 = (1n << 255n) - 19n;
+let pBls12_381 =
+  0x01ae3a4617c510eac63b05c06ca1493b1a22d9f300f5138f1ef3622fba094800170b5d44300000008508c00000000001n;
+let qBls12_381 =
+  0x12ab655e9a2ca55660b44d1e5c37b00159aa76fed00000010a11800000000001n;
+
+let exampleFields = {
+  pastaFp: Fp,
+  pastaFq: Fq,
+  small: createField(pSmall),
+  babybear: createField(pBabybear),
+  goldilocks: createField(pGoldilocks),
+  f25519: createField(p25519),
+  bls12_381_fp: createField(pBls12_381),
+  bls12_381_fq: createField(qBls12_381),
+};
